Added 24/7 Command

+++ lavalink/player/playerDestroy.js
@@ -1,5 +1,5 @@
 module.exports.run = async(client, player) => {
-client.db.delete(`playerdata_${player.guild}`)    
+if(client.db.has(`247_${player.guild}`)) { client.db.delete(`playerdata_${player.guild}`) }
 const plcheck = client.modules.hasrequest(client, player);
 if(plcheck === true) { await client.modules.resetplayer(client, player) }
 else if(plcheck === false) { return; }
